alsa-topo: switch to conf model for the graph
The data model which reflects directly the conf file
domain is more appropriate for the graph modelling
as it is based on a one-to-many relations, more flexible
then a simple tree read from the tplg binary file.

The tplg binary file is now converted into the conf
domain on reading.

Signed-off-by: Marcin Maka <marcin.maka@linux.intel.com>