topo-ui: add generic interface between graph ui and alsa topology
Now the topology graph UI plug-in is able to render topologies
using a generic interface. It is completely detached from the
specific Alsa Topology implementation. It provides extension point
for topology providers. So it is possible to independently develop
and deploy plug-ins for different kinds of topologies.

The Alsa Topology plugin is refactored to contribute to the extension
point and implements the ITopoFactory interface. It is dynamically
discovered by the Topology UI plug-in as a handler for tplg files.

The next step is to rename org.sofproject.alsa.topo.ui into
org.sofproject.topo.ui as its no longer Alsa topology specific.

Signed-off-by: Marcin Maka <marcin.maka@linux.intel.com>