topo: add isChanged to a node attribute interface
Returns true if the attribute has been changed from
the default value.
It may be used by ui in future to highlight changed
attributes.
May be also useful to reduce output and skip the default
optional attributes while serializing the topology
(currently all the attributes are present with their
default values if not set).

Signed-off-by: Marcin Maka <marcin.maka@linux.intel.com>